unidash/backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# HTTP Client & WebSocket
httpx==0.27.0
websockets==13.1

# Database & ORM
sqlalchemy==2.0.36
alembic==1.14.0

# Pydantic & Settings
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Security - JWT & Password Hashing
pyjwt==2.10.1
cryptography==44.0.0
argon2-cffi==23.1.0
passlib[bcrypt]==1.7.4

# Docker Client
docker==7.0.0
urllib3<2.0.0  # Fix for docker socket support
requests<2.32.0  # Compatibility with docker socket

# Type Generation
pydantic-to-typescript==1.0.10

# Utilities
python-dotenv==1.0.0
watchdog==3.0.0

# Logging & Monitoring
structlog==24.1.0
